LayoutTests/imported/w3c/web-platform-tests/webrtc/protocol/crypto-suite.https-expected.txt


PASS tlsVersion is acceptable on data-only
PASS tlsVersion is acceptable on video-only
FAIL dtlsCipher is acceptable on data-only assert_true: expected true got false
FAIL dtlsCipher is acceptable on video-only assert_true: expected true got false
PASS srtpCipher is acceptable on data-only
PASS srtpCipher is acceptable on video-only
FAIL tlsGroup is acceptable on data-only assert_true: Value present: expected true got false
FAIL tlsGroup is acceptable on video-only assert_true: Value present: expected true got false
